[services][feature-flags] Set up the logger
Summary: Use tracing as a logging library - the same as is used in other services.
Added a couple of dependencies with the same versions as used in `backup` service.

Depends on D6663

Test Plan: Run the service and check if log message is displayed.
Introduce two new logs with debug and warn severity and check if only warn is being displayed.
Set `RUST_LOG` env variable to debug and later to warn and check if the logging filtering behaves correctly.

Reviewers: michal, bartek, jon, varun