[keyserver] Use Platform instead of a DeviceType for sending notifs
Summary:
We want to be able to send web notifications but currently `send.js` is typed using `DeviceType` which doesn't include `web`. This diff changes it to use the `Platform` type and changes some names to fit better. There should be no behaviour changes introduced in this diff.

Test Plan:
Create a new account on iOS and check if notifications still work.

Reviewers: tomek, atul, kamil

Subscribers: