[keyserver/ web] Handle device_token on web
Summary: We need to store `device_token` on web so we can save push service endpoints and auth info. This diff make `deviceToken` a proper field in web redux and enables the web app to set it's device token on keyserver.

Test Plan:
Tested with the next diffs in stack that set the `device_token`:
- check that the `device_token` is correctly set in the database and in the redux on web
- check that after logging out the `device_token` is assigned to an anonymous cookie
- check that after another user logs in they get the device's `device_token`

Reviewers: tomek, atul, kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6815