[keyserver][lib] Add COMM_WALLETCONNET_KEY to Webpack build
Summary:
This is similar to how we handle `COMM_ALCHEMY_KEY`. It will be exposed to the Webpack build through the `Dockerfile`, and additionally it will be exposed to `keyserver`. (We don't need it there though.)

This addresses [ENG-2969](https://linear.app/comm/issue/ENG-2969/pass-in-walletconnect-api-key-to-landing-and-web)

Test Plan: Will test in combination with following diffs

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6670