[lib][landing][web] Move Webpack to lib
Summary:
The `webpack` NPM module is imported from `lib` only, so it makes sense for it to be there.

I considered leaving it in `web` and `landing` as well since I suspected perhaps `webpack-dev-server` or `webpack-cli` need it to be a peer dependency. But I didn't see any issues running the standard workflows (see test plan) so I figured it wasn't necessary.

Test Plan: Ran `yarn dev` and `yarn prod` in both `web` and `landing`. Ran the web app and landing page and played around

Reviewers: atul

Subscribers: