[desktop] Continue checking for updates even after an exception
Summary:
[ENG-2923](https://linear.app/comm/issue/ENG-2923)
If `checkForUpdates` fails because of lack of the internet connection `update-not-available` won't run and the timer for checking updated won't be restarted. After this diff check for updates will be done continuously and only stopped between `update-available` and `update-downloaded` events (so the update won't downloaded multiple times).

Test Plan: Add some logging and check that `checkForUpdates` is called correctly, even in case of no network connection. Try updating to a new version.

Reviewers: tomek, atul, kamil, przemek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6683