[desktop] Ignore exceptions about being offline
Summary:
[ENG-2923](https://linear.app/comm/issue/ENG-2923)
If we call `autoUpdater.checkForUpdates()` while we are offline there will be an exception thrown. Unfortunately we can't just `try{}catch{}` but we can catch all exceptions and swallow all exceptions with a message about being offline. We already notify the user about being offline with the DISCONNECTED bar. If it's some other exceptions we can call `dialog.showErrorBox` to get the previous behaviour.

Test Plan: Check if `autoUpdater.checkForUpdates()` is ignored correctly. Check if other exceptions are displayed correctly.

Reviewers: tomek, atul, kamil, przemek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6684