Add .js suffix to all file imports
Summary:
Full context in [ENG-2982](https://linear.app/comm/issue/ENG-2982/specify-file-extension-in-all-imports), but basically Node 19 and Webpack 5 want you to explicitly specify the suffix.

To make this diff I ran a codemod script, which I'm not committing, but is documented in the task.

Test Plan: The way I ran the codemod script is documented in the task. I also tested this diff and the rest of the stack by running `web`, `landing`, `keyserver`, and `native` all together and making sure it still worked

Reviewers: atul, tomek

Differential Revision: https://phab.comm.dev/D6692