Specify removeFullPathAutoPrefix to assets-webpack-plugin
Summary:
After updating to Webpack 5, I found that `assets.json` had an `auto` prefix attached to the paths. I found [this option](https://github.com/ztoben/assets-webpack-plugin#removefullpathautoprefix) "undid" that behavior, but I needed to update my version of `assets-webpack-plugin` to have access to it.

Note that I recently updated this package in D6682, but at the time could not upgrade to the 6.x branch as it required Webpack 5.

Depends on D6699

Test Plan: Confirm that Node is able to return website. (Without this diff it prints an error)

Reviewers: atul, tomek

Differential Revision: https://phab.comm.dev/D6700