[web] Update to react-switch@7.0.0
Summary:
This fixes [ENG-2987](https://linear.app/comm/issue/ENG-2987/unable-to-open-calendar-filters-on-prod), which broke calendar filters on prod.

What happened is similar to D6704. Basically the old version of `react-switch` was packaged as CommonJS, but exported an object with an `__esModule` property. When imported from an ESM context, Webpack 4 would handle this as a special case and allow access to the default export as if it was an ESM export, but Webpack 5 removed that "special case" logic.

Test Plan:
1. Confirm calendar filters open on web again by testing web build in dev mode
2. Review `react-switch`'s [CHANGELOG](https://github.com/markusenglund/react-switch/blob/master/CHANGELOG.md) file and confirm there are no breaking changes that affect us

Reviewers: inka, tomek, atul

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D6711