[native] support operations failure in error handling
Summary:
Follow up to D5672.
If there will be an error while processing store operations we want to note this in secure store and kill app, after opening again database will be recreated.

The problem is that operation failure might occur when migration/creating database works properly - that being said after error in reducer and entry this in `SecureStore`, user opens app and `QueryExecutor` is properly created, which results in setting database status as workable.
This diff resolves that problem by adding one more possible state.

Test Plan: Test if this not break previous functionality of error handling. Check if calling `reportDBOperationsFailure` set value in `SecureStore`

Reviewers: marcin, bartek, inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6712