[web] Remove navigation panel from LeftLayoutAside
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
The `AppSwitcher` will be in a top bar, and `SettingsSwitcher` has to be moved to main content, because of how we use the grid.
So For now I'm removing them from the LeftLayoutAside.

Test Plan: Run web app, check that in inbox and settinngs tabs the switchers are not present

Reviewers: przemek, michal, bartek

Reviewed By: bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6344