[web] Remove LeftLayoutAside
Summary:
Since D6344 the `LeftLayoutAside` is obsolete. I'm removing it, and putting the `CommunityPicker` directly in the `App` component.
This also calls for removing the `left-layout-aside.css`, which I'll do in the next diff, as it requires some code to be moved somewhere else.

Test Plan: Run the web app, check that the sidebar displayes correctly.

Reviewers: przemek, michal, bartek, tomek

Reviewed By: bartek, tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6336