[web] Add a selector fetching thread IDs according to COMMUNITIES calendar filter
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
Since D6645 the calendar filters reducer reacts to `updateCalendarCommunityFilter` and `clearCalendarCommunityFilter` actions. We want to be able to extract the thread IDs based on the filters these actions have set.

Test Plan:
Dispatch `updateCalendarCommunityFilter` and `clearCalendarCommunityFilter` actions, log the value returned by `filteredCommunityThreadIDsSelector`. Check that it is correct. Check that it is only
possible to add an event for a chat in the selected community (in the web app) - this is a result of updating `onScreenThreadInfos` that is used by `ThreadPickerModal`

Reviewers: kamil, tomek

Reviewed By: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6461