[web] Add to calendar filters reducer logic for handling the COMMUNITY filters
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
In D6457 I added architectural changes that allow for a new type of calendar filter. In this diff I add logic to the web reducer, that will handle setting this state.
This used to be diff D6458, but I messed up somethin in my local git, and can't get that commit back, sorry.

Test Plan: Dispatch updateCalendarCommunityFilter and clearCalendarCommunityFilter actions, check in redux tools that they succeed, and redux state changes as expected.

Reviewers: kamil, tomek

Reviewed By: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6645