[web] Highlight drawer items in calendar tab
Summary:
issue: https://linear.app/comm/issue/ENG-2839/highlight-drawer-item-in-the-calendar-tab
We want the drawer item corresponding to the picked community to be highlighted (in calendar tab). `isActive` property of `handled` object is what defines that behaviour - drawer item based on it
applies the highlight or not.

{F369877}

Test Plan: Run web app, checked that drawer items highlight propelry, and when no comunity is picked, no item is highlighted.

Reviewers: kamil, michal

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6661