[web] Add logic for displaying only chats from a given community in calendar filters
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
Add to `FilterPanel` logic, that will allow to display only chats of a given community in filters.

Test Plan:
Checked that when a community is chosen it is not possible to add an event for a chat from a different community - ThreadPickedModal displays only chats from this community. Checked that only chats fro the give community show up in the filter panel.
Check that the square next to the "Your chats" button is filled when all chats from the picked community are selected. Not only when all chats the user is a member of are selected.

Reviewers: kamil, tomek

Reviewed By: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6462