[web] Fix drawer higlight glitch
Summary:
In D6292 I made drawer items react based on the current tab, by having them fetch a `Handler` component type and get the right `handler` for the selected tab. In D6389 I added a `isActive` field to
the
`handler` object they obtain, that is used for determining which item is supposed to be higlighted. Then in D6363 I decided to have the drawer render two drawer item lists, and have one of them be
hidden, based on the current tab (two, becuse we have the same behaviour for all tabs but the Calendar tab). This means, that the items don't need to change the handler they use, because the list of
items is changed instead.  Instead, they should get a `Handler` type from their parent, and always use the same one.

Not doing this was causing glitches when switching tabs

{F370919}

Test Plan:
Rub web app, check that drawer items in different tabs are highlighted properly, and there is no glitching.

{F370920}

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6668