[web] Style the community drawer
Summary:
issue: https://linear.app/comm/issue/ENG-2671/style-community-drawer-on-web
Styling the community drawer (only the drawer for now, the top bar and other elements changed during the drawer tasks will be styled in a separate diff).
I made the header 4px taller for now, because until there is a navigation state info in the main content's header (see https://linear.app/comm/issue/ENG-2740/add-navigation-state-info-to-the-top-bar,
https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=7780%3A136895&t=xSs7qrxIaOTjRJxA-4) the drawers header and main contents header bottom borders are almost aligned, and having them be off by 4px looked bad.
I also talked to the design team and we agreed that the purple line next to the active "Inbox"/"All communities"/"Settings" should be 3px, because otherwise it "sinks under" the browsers border (is not well visible).

{F372984}

[[ https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=7780%3A136669&t=xSs7qrxIaOTjRJxA-4 | Designs ]]

Test Plan: Run web app, check that the styles display correctly.

Reviewers: michal, kuba

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6686