[web] Change the name of communityIDFilter field to calendarPickedCommunityID
Summary:
In D6457 I was asked to change this fileds name. Like I expressed there, it would take a lot more time to chaneg this by chagin it in each diff separately, and rebasin the whole stack, so I hope my
doing this in a separate diff is fine.

Test Plan:
Run web app, check that pressing a drawer item in the calendar tab changes available filters and events.

Reviewers:
michal, kamil, kuba

Subscribers: