[web] Fix drawer item not reaching the edge in full screen
Summary:
When a chat was open, and as such highlighted in the drawer, you could see it didn't quite reach the edge. Because the browser has a border, this was not visible when not in the full screen mode.

Test Plan:
Run web app and electorn app. Check that a highlited item touches the left edge of the screen.

Reviewers:
michal

Subscribers: