[keyserver] Adding `tunnelbroker-client` to the keyserver napi-rs dependencies
Summary:
This diff introduces the integration of the shared Tunnelbroker-client Rust library into the keyserver using the napi-rs as a dependency.
We are updating the `Cargo.toml` file for the `rust-node-addon` and updating the Dockerfile to copy the tunnelbroker-client shared library inside the container.

Linear task: [[ https://linear.app/comm/issue/ENG-2729/implement-napi-rs-bridge-in-keyserver-for-the-tunnelbroker-shared | ENG-2729 ]]

Test Plan: All services should be built successfully according to the CI.

Reviewers: varun, bartek

Reviewed By: varun

Subscribers: jon, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6283