[services] Tunnelbroker - Disabling updating session state when session authentication is skipped
Summary:
This diff introduces disabling the session state updates when the session authentication mechanism is skipped by the config flag.
In this case, the sessions are not used, so we don't need to update their state.

Full context: [[ https://linear.app/comm/issue/ENG-2642/disabling-authentication-mechanism-by-the-config-flag-in-tunnelbroker | ENG-2642 ]]
Linear task: [[ https://linear.app/comm/issue/ENG-3022/minor-fixes-while-testing-tunnelbroker-tunnelbroker-shared-client | ENG-3022 ]]

Test Plan: The DynamoDB database field for the session state is not updating when the config flag `sessions.skip_authentication` is used.

Reviewers: jon, bartek

Reviewed By: bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6747