[lib] Increase ENS cache query timeout to 10s
Summary:
This keeps timing out in my dev environment, which is annoying.

Open to making it 10s only in `__DEV__`, but I figured it didn't hurt to do this for prod too... perhaps in situations with bad connectivity, this isn't unusual for Alchemy?

Test Plan: Confirm that ENS names are now loaded on my iOS Simulator while testing

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6748