[lib] Set viewer's username in pending threads
Summary:
Before this diff, `createPendingThread` would include usernames for everybody EXCEPT the viewer.

But we need to know the viewer's username so that if anybody @-tagged the viewer in the original message, then the viewer's username is properly bolded.

(We also may need to fetch an ENS name for the viewer, in case the ENS name was @-tagged.)

To make sure we always have the viewer's username available, we update `createPendingThread` here to take a `LoggedInUserInfo` instead of a `viewerID`.

Depends on D6758

Test Plan: In combination with diffs I haven't submitted yet, I made sure that @-mentions of the viewer get bolded in a pending sidebar

Reviewers: atul

Subscribers: