[lib][native][web] Include current user as member when calling createPendingThread from baseCreatePendingSidebar
Summary:
This diff should be a no-op, as `createPendingThread` currently separates out the viewer.

In the next diff, we'll update `createPendingThread` to require the viewer to be passed in as a member, so that we can always have the viewer's username available to us.

Depends on D6757

Test Plan: In combination with diffs I haven't submitted yet, I made sure that @-mentions of the viewer get bolded in a pending sidebar

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6758