[lib] Add useCreationSideEffectsFunc to MessageSpec
Summary:
This allows us to define a hook that will get called on the client after a message is successfully created. See more context in [ENG-3034](https://linear.app/comm/issue/ENG-3034/add-usecreationsideeffectsfunc-to-message-specs).

Note that for each `MessageType` that needs this, a callsite will have to be manually added. That's because we handle message creation differently for each composable type.

Depends on D6802

Test Plan:
I tested this stack with the following steps on both `web` and `native`:

1. Make sure users get auto-added to a sidebar when @-mentioned
2. Test mentions typeahead and make sure it includes all members of parent
3. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure they are included in the sidebar creation message
4. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure their name is bolded from the very start of when the message is visible on-screen

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6803