[native] Emit artificial onSelectionChange alongside onChangeText in ClearableTextInput
Summary: When clearing the input in `ClearableTextInput`, we construct an artificial `onChangeText` event. This confuses `SelectableTextInput`, which expects that `onChangeText` events should always be accompanied by `onSelectionChange` events.

As a result of this interaction, a bug occurs because `ClearableTextInput` defers emiting an `onUpdateSyncedSelectionData` event until it receives the corresponding `onSelectionChange`.

This diff addresses the issues by adding an artificial `onSelectionChange` event alongside the artificial `onChangeText` event. On iOS `onSelectionChange` comes first, but on Android it comes last.

Test Plan: Before this change, I had noticed a bug where the @-mention typeahead would not clear after text input cleared. After this change the issue was gone.

Reviewers: tomek, atul

Subscribers: