[native] Create the full screen media gallery component
Summary:
Here, we introduce the full screen media gallery on native that displays all of the media for the thread. This comes with rendering the ThreadSettingsMediaGallery wtih additional props, and also
creating a FilterBar that is used to be able to filter out if we want to view all media, just photos, or just videos. As mentioned in the designs, we don't scroll back to the top automatically when a filter
is selected, but instead we maintain the position.

Linear: https://linear.app/comm/issue/ENG-2929/create-the-full-screen-media-gallery-component-on-native

Depends on D6718

Test Plan: Check the full screen media gallery on the Android emulator, iOS emulator and my physical iPhone

Reviewers: atul, tomek, ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6720