[lib][native][web] Include members of sidebar parent in mentions typeahead
Summary:
This addresses [ENG-3039](https://linear.app/comm/issue/ENG-3039/include-parent-members-in-sidebar-mentions-typeahead). We want to always show all members of the parent in a sidebar's @-mentions typeahead. (Note that members of a sidebar must also be members of its parent.)

Depends on D6807

Test Plan:
I tested this stack with the following steps on both `web` and `native`:

1. Make sure users get auto-added to a sidebar when @-mentioned
2. Test mentions typeahead and make sure it includes all members of parent
3. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure they are included in the sidebar creation message
4. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure their name is bolded from the very start of when the message is visible on-screen

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6808