[native] Patch baseThreadInfo in MessageListContainer with @-mentions in initial sidebar message
Summary:
If a user starts a sidebar by replying to a message with a new text message, and @-mentions somebody in the text message, then that person gets added to the sidebar when it is created.

However, we have a weird visual behavior today where the person's name doesn't appear bolded while the thread is being created. That's because the pending `ThreadInfo` that the `MessageListContainer` is using isn't being updated to include the @-mentions.

This diff adds some plumbing to make this possible.

Depends on D6809

Test Plan:
I tested this stack with the following steps on both `web` and `native`:

1. Make sure users get auto-added to a sidebar when @-mentioned
2. Test mentions typeahead and make sure it includes all members of parent
3. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure they are included in the sidebar creation message
4. If a user is @-mentioned in a text message that triggers the creation of a sidebar, then make sure their name is bolded from the very start of when the message is visible on-screen

Reviewers: atul