[web] Create a new modal for the media gallery on web
Summary: The web client needs a modal to contain all of the shared media for a specific thread. Much of the logic here is the same that is in the native counterpart, D6659.

Linear: https://linear.app/comm/issue/ENG-2878/create-a-new-modal-for-the-media-gallery-on-web

Depends on D6720

Test Plan: The next diff will add the media modal to the thread menu to click through into, but I will show a video of what the modal looks like on `web` below.

Reviewers: atul, ginsu, tomek

Subscribers: