[lib] Add sidebarCreation flag to text and media MessageDatas
Summary:
This flag won't be passed on through to the `RawMessageInfo`. Instead, it's used by `keyserver` code that calls `createMessages` to be able to affect notification behavior deeper in the callstack.

More details in [ENG-3041](https://linear.app/comm/issue/ENG-3041/stop-sending-two-notifs-for-sidebar-creation).

Depends on D6795

Test Plan: I tested this whole stack as follows:

1. Create a sidebar by sending a text message on web
2. Create a sidebar by sending a media message on web
3. Create a sidebar by sending a text message on native
4. Create a sidebar by sending a media message on native

In all of these cases, I confirmed that only one notif was delivered (the notif for the text/media message was skipped)

Reviewers: atul

Subscribers: