[lib] Disable text/media notif when starting sidebar
Summary: More context in [ENG-3041](https://linear.app/comm/issue/ENG-3041/stop-sending-two-notifs-for-sidebar-creation). Today we send two notifs when a sidebar is created; this diff makes it so we only send one notif.

Depends on D6797

Test Plan: I tested this whole stack as follows:

1. Create a sidebar by sending a text message on web
2. Create a sidebar by sending a media message on web
3. Create a sidebar by sending a text message on native
4. Create a sidebar by sending a media message on native

In all of these cases, I confirmed that only one notif was delivered (the notif for the text/media message was skipped)

Reviewers: atul

Subscribers: