Implement native method on Android to get notifications olm account storage path
Summary: This differential extends PlatformSpecificTools.java with method to retrieve notifications olm account path. This functionality requires main application context to be accessible in PlatformSpecificTools. Normally we would pass it in constructor or as method argument, but this
class is designed to be static method only class callable from C++. Therefore we need a "static" way of accessing main application context here. This approach is described here: https://stackoverflow.com/questions/2002288/static-way-to-get-context-in-android. Firther diffs will extend
C++ header to include this method.

Test Plan: Call this method from CommNotificationsHandler on each message and log path to the console.

Reviewers: bartek

Subscribers: